+++ integration-tests/tests/test_cases_runtime.rs
@@ -2,7 +2,8 @@
 mod test {
     use near_chain_configs::Genesis;
     use near_primitives::state_record::StateRecord;
-    use nearcore::config::{GenesisExt, TESTING_INIT_BALANCE};
+    use nearcore::config::TESTING_INIT_BALANCE;
+    use testlib::nearcore_test_utils::GenesisExt;
     use testlib::node::RuntimeNode;
     use testlib::runtime_utils::{add_test_contract, alice_account, bob_account};
     #[cfg(feature = "protocol_feature_evm")]
