+++ nearcore/tests/sync_nodes.rs
@@ -21,8 +21,9 @@ use near_primitives::types::validator_stake::ValidatorStake;
 use near_primitives::types::{BlockHeightDelta, EpochId};
 use near_primitives::validator_signer::{InMemoryValidatorSigner, ValidatorSigner};
 use near_primitives::version::PROTOCOL_VERSION;
-use nearcore::config::{GenesisExt, TESTING_INIT_STAKE};
-use nearcore::{load_test_config, start_with_config, NearConfig};
+use nearcore::config::TESTING_INIT_STAKE;
+use nearcore::{start_with_config, NearConfig};
+use testlib::nearcore_test_utils::{load_test_config, GenesisExt};
 use testlib::{genesis_block, test_helpers::heavy_test};
 
 // This assumes that there is no height skipped. Otherwise epoch hash calculation will be wrong.
