+++ neard/Cargo.toml
@@ -28,6 +28,7 @@ futures = "0.3"
 nearcore = { path = "../nearcore" }
 near-primitives = { path = "../core/primitives" }
 near-performance-metrics = { path = "../utils/near-performance-metrics" }
+testlib = { path = "../test-utils/testlib" }
 
 [features]
 default = ["json_rpc"]
