+++ test-utils/testlib/src/runtime_utils.rs
@@ -1,5 +1,6 @@
 use byteorder::{ByteOrder, LittleEndian};
 
+use crate::nearcore_test_utils::GenesisExt;
 use near_chain_configs::Genesis;
 use near_primitives::account::Account;
 use near_primitives::hash::{hash, CryptoHash};
@@ -7,7 +8,6 @@ use near_primitives::state_record::{state_record_to_account_id, StateRecord};
 use near_primitives::types::{AccountId, StateRoot};
 use near_store::test_utils::create_tries;
 use near_store::{ShardTries, TrieUpdate};
-use nearcore::config::GenesisExt;
 use node_runtime::{state_viewer::TrieViewer, Runtime};
 
 use std::collections::HashSet;
